Android sustained performance mode updates/fixes
* Imagine: Remove UserActivityFaker since it only worked reliably on a handful of Google devices
* Imagine: Add a sustained performance mode using a thread running no-op instructions, only enabled on ARMv7 devices
* Imagine: Add Base::sustainedPerformanceModeType() to determine if the device has native support or simulates it with a no-op thread
* EmuFramework: Fix not checking optionSustainedPerformanceMode before calling Base::setSustainedPerformanceMode()
* EmuFramework: Disable sustained performance mode by default if using a no-op thread

+++ imagine/tests/FrameRateTest/src/main/tests.cc
@@ -148,9 +148,9 @@ void TestFramework::frameUpdate(Gfx::RendererTask &rTask, Base::Window &win, Bas
 		if(frames && frames % 4 == 0)
 		{
 			string_printf(statsStr, "Process: %02lums (%02ums)\nPresent: %02lums (%02ums)",
-				std::chrono::duration_cast<IG::Milliseconds>(lastFramePresentTime.atWinPresent - lastFramePresentTime.atOnFrame).count(),
+				(unsigned long)std::chrono::duration_cast<IG::Milliseconds>(lastFramePresentTime.atWinPresent - lastFramePresentTime.atOnFrame).count(),
 				lostFrameProcessTime,
-				std::chrono::duration_cast<IG::Milliseconds>(lastFramePresentTime.atWinPresentEnd - lastFramePresentTime.atWinPresent).count(),
+				(unsigned long)std::chrono::duration_cast<IG::Milliseconds>(lastFramePresentTime.atWinPresentEnd - lastFramePresentTime.atWinPresent).count(),
 				lostFramePresentTime);
 			updatedFrameStats = true;
 		}
